+++ src/main/java/com/msy/wallet/service/WalletService.java
@@ -6,5 +6,5 @@ import com.msy.wallet.model.User;
 
 public interface WalletService {
     User getBalance(Long userId) throws WalletServiceException;;
-    Transaction addMoney(Long userId, Double amount);
+    Transaction addMoney(Long userId, Long amount);
 }
